move post to data folder and update university to college
